--- app/src/main/java/com/smilehunter/ablebody/domain/OrderItemUseCase.kt
@@ -1,21 +1,0 @@
-package com.smilehunter.ablebody.domain
-
-import com.smilehunter.ablebody.network.model.request.AddOrderListRequest
-import com.smilehunter.ablebody.data.repository.OrderManagementRepository
-import com.smilehunter.ablebody.network.di.AbleBodyDispatcher.IO
-import com.smilehunter.ablebody.network.di.Dispatcher
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.withContext
-import javax.inject.Inject
-
-class OrderItemUseCase @Inject constructor(
-    @Dispatcher(IO) private val ioDispatcher: CoroutineDispatcher,
-    private val orderManagementRepository: OrderManagementRepository
-) {
-
-    suspend operator fun invoke(
-        addOrderListRequest: AddOrderListRequest
-    ): String = withContext(ioDispatcher) {
-        orderManagementRepository.orderItem(addOrderListRequest).data!!.orderId
-    }
-}
