+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/EditProfileUseCase.kt
@@ -1,4 +1,4 @@
-package com.smilehunter.ablebody.domain
+package com.smilehunter.ablebody.domain.usecase
 
 import com.smilehunter.ablebody.network.model.request.EditProfile
 import com.smilehunter.ablebody.data.repository.UserRepository
