+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetCouponListUseCase.kt
@@ -1,16 +1,16 @@
-package com.smilehunter.ablebody.domain
+package com.smilehunter.ablebody.domain.usecase
 
-import com.smilehunter.ablebody.network.model.response.GetCouponBagsResponseData
 import com.smilehunter.ablebody.data.repository.UserRepository
 import com.smilehunter.ablebody.model.CouponData
-import com.smilehunter.ablebody.network.di.AbleBodyDispatcher.*
+import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
 import com.smilehunter.ablebody.network.di.Dispatcher
+import com.smilehunter.ablebody.network.model.response.GetCouponBagsResponseData
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class GetCouponListUseCase @Inject constructor(
-    @Dispatcher(IO) private val ioDispatcher: CoroutineDispatcher,
+    @Dispatcher(AbleBodyDispatcher.IO) private val ioDispatcher: CoroutineDispatcher,
     private val userRepository: UserRepository
 ) {
