+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetCreatorDetailDataListUseCase.kt
@@ -1,9 +1,9 @@
-package com.smilehunter.ablebody.domain
+package com.smilehunter.ablebody.domain.usecase
 
 import com.smilehunter.ablebody.data.model.ItemChildCategory
-import com.smilehunter.ablebody.network.model.response.CreatorDetailResponseData
 import com.smilehunter.ablebody.data.repository.CreatorDetailRepository
 import com.smilehunter.ablebody.model.CreatorDetailData
+import com.smilehunter.ablebody.network.model.response.CreatorDetailResponseData
 import javax.inject.Inject
 
 class GetCreatorDetailDataListUseCase @Inject constructor(
