+++ app/src/main/java/com/smilehunter/ablebody/domain/usecase/GetRecommendKeywordUseCase.kt
@@ -1,4 +1,4 @@
-package com.smilehunter.ablebody.domain
+package com.smilehunter.ablebody.domain.usecase
 
 import com.smilehunter.ablebody.data.repository.SearchRepository
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
@@ -7,14 +7,14 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
-class DeleteSearchHistoryUseCase @Inject constructor(
+class GetRecommendKeywordUseCase @Inject constructor(
     @Dispatcher(AbleBodyDispatcher.IO) private val ioDispatcher: CoroutineDispatcher,
     private val searchRepository: SearchRepository
 ) {
 
-    suspend operator fun invoke() {
-        withContext(ioDispatcher) {
-            searchRepository.deleteAllSearchHistory()
+    suspend operator fun invoke(): List<String> {
+        return withContext(ioDispatcher) {
+            searchRepository.uniSearch("").data!!.recommendKeyword.creator
         }
     }
 }
