+++ app/src/main/java/com/smilehunter/ablebody/presentation/home/brand/BrandViewModel.kt
@@ -6,7 +6,7 @@ import com.smilehunter.ablebody.data.model.ItemGender
 import com.smilehunter.ablebody.data.model.SortingMethod
 import com.smilehunter.ablebody.data.result.Result
 import com.smilehunter.ablebody.data.result.asResult
-import com.smilehunter.ablebody.domain.GetBrandListUseCase
+import com.smilehunter.ablebody.domain.usecase.GetBrandListUseCase
 import com.smilehunter.ablebody.presentation.home.brand.data.BrandListResultUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
