+++ app/src/main/java/com/smilehunter/ablebody/presentation/item_detail/ItemDetailViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.viewModelScope
 import com.smilehunter.ablebody.data.repository.BookmarkRepository
 import com.smilehunter.ablebody.data.result.Result
 import com.smilehunter.ablebody.data.result.asResult
-import com.smilehunter.ablebody.domain.GetItemOptionListUseCase
+import com.smilehunter.ablebody.domain.usecase.GetItemOptionListUseCase
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher.IO
 import com.smilehunter.ablebody.network.di.Dispatcher
 import com.smilehunter.ablebody.presentation.item_detail.data.ItemDetailUiState
