+++ app/src/main/java/com/smilehunter/ablebody/presentation/like_list/LikeListViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.smilehunter.ablebody.data.result.Result
 import com.smilehunter.ablebody.data.result.asResult
-import com.smilehunter.ablebody.domain.GetLikeListUseCase
+import com.smilehunter.ablebody.domain.usecase.GetLikeListUseCase
 import com.smilehunter.ablebody.model.LikedLocations
 import com.smilehunter.ablebody.presentation.like_list.data.LikeListUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
