+++ app/src/main/java/com/smilehunter/ablebody/presentation/my/MyViewModel.kt
@@ -9,9 +9,9 @@ import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import com.smilehunter.ablebody.data.repository.ManageRepository
-import com.smilehunter.ablebody.domain.GetOrderItemListUseCase
-import com.smilehunter.ablebody.domain.GetUserBoardPagerUseCase
-import com.smilehunter.ablebody.domain.GetUserInfoUseCase
+import com.smilehunter.ablebody.domain.usecase.GetOrderItemListUseCase
+import com.smilehunter.ablebody.domain.usecase.GetUserBoardPagerUseCase
+import com.smilehunter.ablebody.domain.usecase.GetUserInfoUseCase
 import com.smilehunter.ablebody.model.OrderItemData
 import com.smilehunter.ablebody.model.UserBoardData
 import com.smilehunter.ablebody.model.UserInfoData
