+++ app/src/main/java/com/smilehunter/ablebody/presentation/my/coupon/CouponViewModel.kt
@@ -1,28 +1,20 @@
 package com.smilehunter.ablebody.presentation.my.coupon
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.smilehunter.ablebody.data.repository.UserRepository
-import com.smilehunter.ablebody.domain.AddCouponUseCase
-import com.smilehunter.ablebody.domain.CouponStatus
-import com.smilehunter.ablebody.domain.GetCouponListUseCase
-import com.smilehunter.ablebody.domain.GetOrderItemListUseCase
-import com.smilehunter.ablebody.domain.GetUserBoardPagerUseCase
-import com.smilehunter.ablebody.domain.GetUserInfoUseCase
+import com.smilehunter.ablebody.domain.usecase.AddCouponUseCase
+import com.smilehunter.ablebody.domain.usecase.CouponStatus
+import com.smilehunter.ablebody.domain.usecase.GetCouponListUseCase
 import com.smilehunter.ablebody.model.CouponData
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
 import com.smilehunter.ablebody.network.di.Dispatcher
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
-import kotlin.coroutines.suspendCoroutine
 
 @HiltViewModel
 class CouponViewModel @Inject constructor(
