+++ app/src/main/java/com/smilehunter/ablebody/presentation/my/myInfo/MyInfoViewModel.kt
@@ -6,27 +6,14 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import androidx.paging.PagingData
-import androidx.paging.cachedIn
-import com.smilehunter.ablebody.data.repository.OnboardingRepository
 import com.smilehunter.ablebody.data.repository.TokenRepository
 import com.smilehunter.ablebody.data.repository.UserRepository
-import com.smilehunter.ablebody.domain.AddCouponUseCase
-import com.smilehunter.ablebody.domain.GetCouponListUseCase
-import com.smilehunter.ablebody.domain.GetOrderItemListUseCase
-import com.smilehunter.ablebody.domain.GetUserBoardPagerUseCase
-import com.smilehunter.ablebody.domain.GetUserInfoUseCase
-import com.smilehunter.ablebody.model.UserBoardData
+import com.smilehunter.ablebody.domain.usecase.GetUserInfoUseCase
 import com.smilehunter.ablebody.model.UserInfoData
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
 import com.smilehunter.ablebody.network.di.Dispatcher
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asSharedFlow
-import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import javax.inject.Inject
