+++ app/src/main/java/com/smilehunter/ablebody/presentation/my/other/OtherUserViewModel.kt
@@ -1,6 +1,5 @@
 package com.smilehunter.ablebody.presentation.my.other
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
@@ -9,11 +8,11 @@ import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import com.smilehunter.ablebody.data.repository.UserRepository
-import com.smilehunter.ablebody.domain.AddCouponUseCase
-import com.smilehunter.ablebody.domain.GetCouponListUseCase
-import com.smilehunter.ablebody.domain.GetOrderItemListUseCase
-import com.smilehunter.ablebody.domain.GetUserBoardPagerUseCase
-import com.smilehunter.ablebody.domain.GetUserInfoUseCase
+import com.smilehunter.ablebody.domain.usecase.AddCouponUseCase
+import com.smilehunter.ablebody.domain.usecase.GetCouponListUseCase
+import com.smilehunter.ablebody.domain.usecase.GetOrderItemListUseCase
+import com.smilehunter.ablebody.domain.usecase.GetUserBoardPagerUseCase
+import com.smilehunter.ablebody.domain.usecase.GetUserInfoUseCase
 import com.smilehunter.ablebody.model.UserBoardData
 import com.smilehunter.ablebody.model.UserInfoData
 import com.smilehunter.ablebody.network.di.AbleBodyDispatcher
@@ -21,10 +20,8 @@ import com.smilehunter.ablebody.network.di.Dispatcher
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.stateIn
