+++ app/src/main/java/com/smilehunter/ablebody/presentation/order_management/OrderManagementViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.viewModelScope
 import com.smilehunter.ablebody.data.repository.OrderManagementRepository
 import com.smilehunter.ablebody.data.result.Result
 import com.smilehunter.ablebody.data.result.asResult
-import com.smilehunter.ablebody.domain.GetDeliveryTrackingNumberUseCase
-import com.smilehunter.ablebody.domain.GetOrderItemListUseCase
+import com.smilehunter.ablebody.domain.usecase.GetDeliveryTrackingNumberUseCase
+import com.smilehunter.ablebody.domain.usecase.GetOrderItemListUseCase
 import com.smilehunter.ablebody.presentation.order_management.data.DeliveryTrackingUiState
 import com.smilehunter.ablebody.presentation.order_management.data.OrderItemUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
