+++ app/src/main/java/com/smilehunter/ablebody/presentation/payment/PaymentViewModel.kt
@@ -6,12 +6,12 @@ import androidx.lifecycle.viewModelScope
 import com.smilehunter.ablebody.network.model.request.AddOrderListRequest
 import com.smilehunter.ablebody.data.result.Result
 import com.smilehunter.ablebody.data.result.asResult
-import com.smilehunter.ablebody.domain.ConfirmPaymentUseCase
-import com.smilehunter.ablebody.domain.GetCouponListUseCase
-import com.smilehunter.ablebody.domain.GetMyDeliveryAddressUseCase
-import com.smilehunter.ablebody.domain.GetUserInfoUseCase
-import com.smilehunter.ablebody.domain.HandlePaymentFailureUseCase
-import com.smilehunter.ablebody.domain.OrderItemUseCase
+import com.smilehunter.ablebody.domain.usecase.ConfirmPaymentUseCase
+import com.smilehunter.ablebody.domain.usecase.GetCouponListUseCase
+import com.smilehunter.ablebody.domain.usecase.GetMyDeliveryAddressUseCase
+import com.smilehunter.ablebody.domain.usecase.GetUserInfoUseCase
+import com.smilehunter.ablebody.domain.usecase.HandlePaymentFailureUseCase
+import com.smilehunter.ablebody.domain.usecase.OrderItemUseCase
 import com.smilehunter.ablebody.model.CouponData
 import com.smilehunter.ablebody.model.UserInfoData
 import com.smilehunter.ablebody.presentation.payment.data.CouponBagsUiState
