+++ app/src/main/java/com/smilehunter/ablebody/presentation/receipt/ReceiptViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.smilehunter.ablebody.data.result.Result
 import com.smilehunter.ablebody.data.result.asResult
-import com.smilehunter.ablebody.domain.GetReceiptUseCase
+import com.smilehunter.ablebody.domain.usecase.GetReceiptUseCase
 import com.smilehunter.ablebody.presentation.receipt.data.ReceiptUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
