+++ app/src/main/java/com/smilehunter/ablebody/presentation/search/SearchViewModel.kt
@@ -13,13 +13,13 @@ import com.smilehunter.ablebody.data.model.PersonHeightFilterType
 import com.smilehunter.ablebody.data.model.SortingMethod
 import com.smilehunter.ablebody.data.result.Result
 import com.smilehunter.ablebody.data.result.asResult
-import com.smilehunter.ablebody.domain.CodyItemPagerUseCase
-import com.smilehunter.ablebody.domain.CodyPagingSourceData
-import com.smilehunter.ablebody.domain.DeleteSearchHistoryUseCase
-import com.smilehunter.ablebody.domain.GetRecommendKeywordUseCase
-import com.smilehunter.ablebody.domain.GetSearchHistoryUseCase
-import com.smilehunter.ablebody.domain.ProductItemPagerUseCase
-import com.smilehunter.ablebody.domain.ProductItemPagingSourceData
+import com.smilehunter.ablebody.domain.usecase.CodyItemPagerUseCase
+import com.smilehunter.ablebody.domain.usecase.CodyPagingSourceData
+import com.smilehunter.ablebody.domain.usecase.DeleteSearchHistoryUseCase
+import com.smilehunter.ablebody.domain.usecase.GetRecommendKeywordUseCase
+import com.smilehunter.ablebody.domain.usecase.GetSearchHistoryUseCase
+import com.smilehunter.ablebody.domain.usecase.ProductItemPagerUseCase
+import com.smilehunter.ablebody.domain.usecase.ProductItemPagingSourceData
 import com.smilehunter.ablebody.model.CodyItemData
 import com.smilehunter.ablebody.model.ProductItemData
 import com.smilehunter.ablebody.model.SearchHistoryQuery
